Remove hashtable
To prevent loops it suffices to look at parent directories.
Therefore simply store device and inode together with each path component.
Due to memory sharing in lists memory requirement will be
O(max_depth + largest dir) instead of O(all dirs).
CPU usage is O(MIN(max_depth, PATH_MAX).

This also changes behavior in that several symlinks to the same directory
will be followed and the directory will be listed multiple times. This
matches the behaviour of posix find and is reasonable because otherwise a
directory (and its contents) might be listed via the symlink only, but not
via the original directory.